Moves Search class to Listbox directory; removes export
