+++ utils/index.js
@@ -2,11 +2,9 @@
  * Select utility function exports.
  */
 import interactiveChildren from '../src/lib/interactiveChildren';
-import Search from '../src/lib/Search';
 import isInstanceOf from '../src/lib/isInstanceOf';
 
 export {
   interactiveChildren,
-  Search,
   isInstanceOf,
 };
